Fixing system history bugs
